PRC-256 : API - PATCH for contact relationship - refactor relationship into its own controller and improve for performance,
