+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/resource/ContactController.kt
@@ -13,7 +13,6 @@ import org.slf4j.LoggerFactory
 import org.springframework.data.domain.Pageable
 import org.springframework.data.domain.Sort.Direction
 import org.springframework.data.web.PageableDefault
-import org.springframework.http.HttpStatus
 import org.springframework.http.MediaType
 import org.springframework.http.ResponseEntity
 import org.springframework.security.access.prepost.PreAuthorize
@@ -24,9 +23,7 @@ import org.springframework.web.bind.annotation.PathVariable
 import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
-import org.springframework.web.bind.annotation.ResponseStatus
 import org.springframework.web.bind.annotation.RestController
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.AddContactRelationshipRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactSearchRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.patch.PatchContactRequest
@@ -133,42 +130,6 @@ class ContactController(
       logger.info("Couldn't find contact with id '{}'", contactId)
       ResponseEntity.notFound().build()
     }
-  }
-
-  @PostMapping("/{contactId}/relationship", consumes = [MediaType.APPLICATION_JSON_VALUE])
-  @Operation(
-    summary = "Add a new contact relationship",
-    description = "Creates a new relationship between the contact and a prisoner.",
-  )
-  @ApiResponses(
-    value = [
-      ApiResponse(
-        responseCode = "201",
-        description = "Created the relationship successfully",
-      ),
-      ApiResponse(
-        responseCode = "400",
-        description = "The request has invalid or missing fields",
-        content = [Content(schema = Schema(implementation = ErrorResponse::class))],
-      ),
-      ApiResponse(
-        responseCode = "404",
-        description = "Could not find the prisoner or contact that this relationship relates to",
-        content = [Content(schema = Schema(implementation = ErrorResponse::class))],
-      ),
-    ],
-  )
-  @PreAuthorize("hasAnyRole('ROLE_CONTACTS_ADMIN')")
-  @ResponseStatus(HttpStatus.CREATED)
-  fun addContactRelationship(
-    @PathVariable("contactId") @Parameter(
-      name = "contactId",
-      description = "The id of the contact",
-      example = "123456",
-    ) contactId: Long,
-    @Valid @RequestBody relationshipRequest: AddContactRelationshipRequest,
-  ) {
-    contactService.addContactRelationship(contactId, relationshipRequest)
   }
 
   @GetMapping("/search")
