+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/resource/RelationshipController.kt
@@ -1,0 +1,68 @@
+package uk.gov.justice.digital.hmpps.hmppscontactsapi.resource
+
+import io.swagger.v3.oas.annotations.Operation
+import io.swagger.v3.oas.annotations.Parameter
+import io.swagger.v3.oas.annotations.media.Content
+import io.swagger.v3.oas.annotations.media.Schema
+import io.swagger.v3.oas.annotations.responses.ApiResponse
+import io.swagger.v3.oas.annotations.responses.ApiResponses
+import io.swagger.v3.oas.annotations.tags.Tag
+import jakarta.validation.Valid
+import org.springframework.http.HttpStatus
+import org.springframework.http.MediaType
+import org.springframework.security.access.prepost.PreAuthorize
+import org.springframework.web.bind.annotation.PathVariable
+import org.springframework.web.bind.annotation.PostMapping
+import org.springframework.web.bind.annotation.RequestBody
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.RestController
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.AddContactRelationshipRequest
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.service.RelationshipService
+import uk.gov.justice.digital.hmpps.hmppscontactsapi.swagger.AuthApiResponses
+import uk.gov.justice.hmpps.kotlin.common.ErrorResponse
+
+@Tag(name = "Contact")
+@RestController
+@RequestMapping(value = ["contact"], produces = [MediaType.APPLICATION_JSON_VALUE])
+@AuthApiResponses
+class RelationshipController(
+  val relationshipService: RelationshipService,
+) {
+
+  @PostMapping("/{contactId}/relationship", consumes = [MediaType.APPLICATION_JSON_VALUE])
+  @Operation(
+    summary = "Add a new contact relationship",
+    description = "Creates a new relationship between the contact and a prisoner.",
+  )
+  @ApiResponses(
+    value = [
+      ApiResponse(
+        responseCode = "201",
+        description = "Created the relationship successfully",
+      ),
+      ApiResponse(
+        responseCode = "400",
+        description = "The request has invalid or missing fields",
+        content = [Content(schema = Schema(implementation = ErrorResponse::class))],
+      ),
+      ApiResponse(
+        responseCode = "404",
+        description = "Could not find the prisoner or contact that this relationship relates to",
+        content = [Content(schema = Schema(implementation = ErrorResponse::class))],
+      ),
+    ],
+  )
+  @PreAuthorize("hasAnyRole('ROLE_CONTACTS_ADMIN')")
+  @ResponseStatus(HttpStatus.CREATED)
+  fun create(
+    @PathVariable("contactId") @Parameter(
+      name = "contactId",
+      description = "The id of the contact",
+      example = "123456",
+    ) contactId: Long,
+    @Valid @RequestBody relationshipRequest: AddContactRelationshipRequest,
+  ) {
+    relationshipService.create(contactId, relationshipRequest)
+  }
+}
