+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppscontactsapi/resource/ContactControllerTest.kt
@@ -5,7 +5,6 @@ import org.junit.jupiter.api.Assertions.assertNotNull
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import org.mockito.Mockito.doNothing
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.verify
 import org.mockito.kotlin.whenever
@@ -18,8 +17,6 @@ import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactAddres
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactEmailDetails
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactIdentityDetails
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.helpers.createContactPhoneNumberDetails
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.AddContactRelationshipRequest
-import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactRelationship
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.ContactSearchRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.CreateContactRequest
 import uk.gov.justice.digital.hmpps.hmppscontactsapi.model.request.EstimatedIsOverEighteen
@@ -142,37 +139,6 @@ class ContactControllerTest {
 
       assertThrows<RuntimeException>("Bang!") {
         controller.getContact(id)
-      }
-    }
-  }
-
-  @Nested
-  inner class AddContactRelationship {
-    private val id = 123456L
-    private val relationship = ContactRelationship(
-      prisonerNumber = "A1234BC",
-      relationshipCode = "MOT",
-      isNextOfKin = true,
-      isEmergencyContact = false,
-      comments = "Foo",
-    )
-    private val request = AddContactRelationshipRequest(relationship, "USER")
-
-    @Test
-    fun `should create a contact relationship successfully`() {
-      doNothing().whenever(contactService).addContactRelationship(id, request)
-
-      controller.addContactRelationship(id, request)
-
-      verify(contactService).addContactRelationship(id, request)
-    }
-
-    @Test
-    fun `should propagate exceptions getting a contact`() {
-      whenever(contactService.addContactRelationship(id, request)).thenThrow(RuntimeException("Bang!"))
-
-      assertThrows<RuntimeException>("Bang!") {
-        controller.addContactRelationship(id, request)
       }
     }
   }
